Langchain_modal/2.ChatModels/chatmodel_openai.py
from langchain_openai import ChatOpenAI
from dotenv import load_dotenv

load_dotenv()

model = ChatOpenAI(model='gpt-4',temperature=0)

result = model.invoke("what is the cpaital of india")

print(result)